moved pins to high-impedance state after programing
Original code was never intended to run along the game.

This might be a non issue, but could create bus conflicts that could stress the Arduino or board in the long run specially on revision 3 and 4 boards. 

Also added a blink to the led after flashing the keys, as a visual cue just after the blink for boot.